extrapolate method to verify signature
