+++ src/server/utils/utils.js
@@ -1,5 +1,6 @@
 import config from '../config';
 import log from '../logger';
+import headerParser from './headerParser';
 import rs from 'jsrsasign';
 
 export const firstLetterUpperCase = (str) => str.charAt(0).toUpperCase() + str.slice(1);
@@ -53,9 +54,27 @@ export const loadPublicKey = () => {
     return publicKey;
   } catch (err) {
       log.error('[KEY LOAD] error when loading the public key', err);
+      return null;
   }
   
   return null;
+}
+
+export const validSignature = (req) => {
+  var isValid = false;
+  try {
+    const signature = headerParser.parseSignature(req);
+    var data = req.body;
+    data = JSON.stringify(data);
+
+    var public_key = req.app.locals.publicKey;
+    isValid = public_key.verify(data, signature) 
+  } catch (err) {
+      log.error('[SIGNATURE CHECK] error when checking the signature of the payload', err);
+      return false;
+  }
+  log.info('The body signature has been decoded: ', isValid);
+  return isValid;
 }
 
 /**
